<ruchir> refactor: move core node-type translators to `expr_translators.rs`
- Extracted `translate_op_expr`, `translate_var`, `translate_const`, and `translate_bool_expr` from `PredicateTranslator` into a split `impl` block in `expr_translators.rs` for better modularity.
- Updated imports and method visibility to support the new structure.
